requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-experimental>=0.0.50
langgraph>=0.0.40
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
openpyxl>=3.1.0